trese

+++ Assets/Leo/Scripts/PlayerStuff/PositionRendering.cs
@@ -7,7 +7,7 @@ public class PositionRendering : MonoBehaviour
     [SerializeField]
     private int sortingOrderBase = 5000;
     [SerializeField]
-    private int offset = 0;
+    private float offset = 0;
     [SerializeField]
     private bool runOnlyOnce = false;
